Application overview
Explanation

This web component was created to help visualise complex surveys datasets, it displays the catalogue variables following the order previously defined by the owner of the data set. The variables catalogue is displayed in the hierarchical order of the group name. The list can also be manipulated to the original order or refined to show subsets of the data.